Change some function names and reorganize them
