+++ Sources/UIViewKit/Previews/IBPreview.swift
@@ -1,0 +1,13 @@
+//
+//  IBPreview.swift
+//  UIViewKit
+//
+//  Created by Blazej SLEBODA on 10/02/2024.
+//
+
+import Foundation
+
+public class IBPreview {
+
+    private init() { }
+}
