+++ Tests/UIViewKitTests/DebugTests/UIViewDebugAllSubviewsTests.swift
@@ -11,7 +11,7 @@ import XCTest
 
 class UIViewDebugTests: XCTestCase {
 
-    typealias SUT = UIViewDebug
+    typealias SUT = IBDebug
 
     func testShow() throws {
         let view = HelloWordView()
